pyproject.toml
[project]
name = "loopbench"
version = "0.1.0"
description = "Harness-first benchmark for closed-loop coding agents"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0.1",
  "rich>=13.7",
  "e2b-code-interpreter>=1.0.0",
]

[project.scripts]
loopbench = "loopbench.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["loopbench*"]
